+++ web/src/features/course/components/course-editor-information.tsx
@@ -23,6 +23,7 @@ import { FormSelect } from "@/shared/components/form/form-select";
 import { FormTextarea } from "@/shared/components/form/form-textarea";
 import { OverlayStatusWrapper } from "@/shared/components/overlay-status-wrapper";
 import { Card, CardContent, CardFooter } from "@/shared/components/shadcn/card";
+import { useAlertDialog } from "@/shared/components/modals/use-alert-dialog";
 import { Form } from "@/shared/components/shadcn/form";
 import {
   MultiSelectOption,
@@ -38,6 +39,11 @@ import {
 } from "../api/course-mutations";
 
 import CategoryCreateModal from "./category-create-modal";
+import { Button } from "@/shared/components/shadcn/button";
+import { mdiArrowLeft } from "@mdi/js";
+import Icon from "@mdi/react";
+import { useNavigate } from "react-router";
+import ReusableAlertDialog from "@/shared/components/modals/reusable-alert-dialog";
 
 /* ------------------------------- Interfaces ------------------------------- */
 interface CourseEditorInformationProps {
@@ -80,6 +86,9 @@ const CourseEditorInformation = forwardRef<
 >(({ course, onComplete }, ref) => {
   const { t } = useTranslation();
   const isEditMode = course !== undefined;
+  const { alertProps, openAlert } = useAlertDialog();
+  const navigate = useNavigate();
+  const informationFormRef = useRef<CourseEditorInformationRef>(null);
 
   /* -------------------------------- Mutations ------------------------------- */
   const createMutation = useCreateCourseMutation();
@@ -170,6 +179,21 @@ const CourseEditorInformation = forwardRef<
   }, [categoriesError, form, t]);
 
   /* -------------------------------- Handlers -------------------------------- */
+
+  const handleCancel = () => {
+    // Check isDirty at the moment of click
+    const hasChanges = form.formState.isDirty;
+
+    if (hasChanges) {
+      openAlert();
+    } else {
+      navigate("/");
+    }
+  };
+
+  const handleReturnToCourses = () => {
+    navigate("/");
+  };
 
   const onSubmit = async (values: CourseBasicInfoFormValues) => {
     try {
@@ -426,24 +450,59 @@ const CourseEditorInformation = forwardRef<
           onCreated={handleCategoryCreated}
         />
       </Card>
-      <div className="flex justify-end mt-4">
-        <FormActions
-          formState={form.formState}
-          formId="course-information-form"
-          submitLabel={
-            isEditMode
-              ? t("common.saveChanges")
-              : t("courseManager.createAndContinue")
-          }
-          submittingLabel={
-            isEditMode
-              ? t("common.saving") + "..."
-              : t("common.creating") + "..."
-          }
-          disableSubmit={mutationError !== undefined}
-        />
+      <div className="bg-white sticky bottom-2 py-4 border-t flex justify-between mt-5 items-center pr-2">
+        <Button
+          disabled
+          className="text-md text-greyscale-border-default"
+          variant="ghost"
+        >
+          &#x276E; &nbsp; &nbsp;
+          {t("common.goPrevious")}
+        </Button>
+        <div className="flex gap-x-5">
+          <Button
+            variant="ghost"
+            className="text-md text-error-surface-default font-bold underline cursor-pointer"
+            onClick={handleCancel}
+          >
+            {t("common.cancel")}
+          </Button>
+
+          <FormActions
+            formState={form.formState}
+            formId="course-information-form"
+            submitLabel={
+              isEditMode
+                ? t("common.saveChanges")
+                : t("courseManager.createAndContinue")
+            }
+            submittingLabel={
+              isEditMode
+                ? t("common.saving") + "..."
+                : t("common.creating") + "..."
+            }
+            disableSubmit={mutationError !== undefined}
+          />
+        </div>
       </div>
       <DevTool control={form.control} />
+
+      {/* Leave confirmation alert */}
+      <ReusableAlertDialog
+        {...alertProps}
+        title={t("courseManager.unsavedChangesTitle")}
+        description={t("courseManager.unsavedChangesMessage")}
+        confirmAction={{
+          label: t("common.leave"),
+          onClick: handleReturnToCourses,
+        }}
+        cancelAction={{
+          label: t("common.stay"),
+          onClick: () => {
+            // Just close the dialog
+          },
+        }}
+      />
     </>
   );
 });
