+++ web/src/features/course/pages/course-editor-page.tsx
@@ -3,13 +3,11 @@ import Icon from "@mdi/react";
 import { useQuery } from "@tanstack/react-query";
 import { useState, useRef } from "react";
 import { useTranslation } from "react-i18next";
-import { useParams, useNavigate } from "react-router-dom";
+import { useParams } from "react-router-dom";
 
 import { ErrorBoundary } from "@/shared/components/error/error-boundary";
 import { ErrorDisplay } from "@/shared/components/error/error-display";
 import { GlobalLoader } from "@/shared/components/global-loader";
-import ReusableAlertDialog from "@/shared/components/modals/reusable-alert-dialog";
-import { useAlertDialog } from "@/shared/components/modals/use-alert-dialog";
 import { PageContainer } from "@/shared/components/page-container";
 import { Button } from "@/shared/components/shadcn/button";
 import { Separator } from "@/shared/components/shadcn/seperator";
@@ -31,8 +29,6 @@ import {
 
 const CourseEditorPage = () => {
   const { t } = useTranslation();
-  const navigate = useNavigate();
-  const { alertProps, openAlert } = useAlertDialog();
 
   // Refs to access form state from child components. Used to prevent navigation with unsaved changes.
   const informationFormRef = useRef<CourseEditorInformationRef>(null);
@@ -167,39 +163,12 @@ const CourseEditorPage = () => {
     return `${t("common.create")} ${t("courseManager.course")}`;
   };
 
-  const handleBack = () => {
-    // Check isDirty at the moment of click
-    const informationDirty = informationFormRef.current?.isDirty() ?? false;
-    const sectionsDirty = sectionsFormRef.current?.isDirty() ?? false;
-
-    const hasChanges = informationDirty || sectionsDirty;
-
-    if (hasChanges) {
-      openAlert();
-    } else {
-      handleReturnToCourses();
-    }
-  };
-
-  const handleReturnToCourses = () => {
-    navigate("/courses");
-  };
-
   return (
     <PageContainer title={getPageTitle()}>
       <div className="flex gap-x-20">
         {/*------------ Left side - Step Navigation ------------*/}
         <div className="w-auto">
           <div className="flex flex-col justify-between items-start gap-4 mb-6">
-            <Button
-              variant="outline"
-              onClick={handleBack}
-              iconPlacement="left"
-              size="sm"
-              icon={() => <Icon path={mdiArrowLeft} size={1} />}
-            >
-              {t("common.back")}
-            </Button>
             <h1 className="text-2xl text-greyscale-text-caption">
               {getPageTitle()}
             </h1>
@@ -220,7 +189,7 @@ const CourseEditorPage = () => {
             ))}
           </div>
           <Separator className="my-6" />
-          <Button className="w-full font-bold" variant="secondary">
+          <Button className="w-full fon" variant="secondary">
             <Icon path={mdiFloppy} size={0.6} /> {t("courseEditor.saveAsDraft")}
           </Button>
         </div>
@@ -233,23 +202,6 @@ const CourseEditorPage = () => {
           <ErrorBoundary>{getSectionComponent()}</ErrorBoundary>
         </div>
       </div>
-
-      {/* Leave confirmation alert */}
-      <ReusableAlertDialog
-        {...alertProps}
-        title={t("courseManager.unsavedChangesTitle")}
-        description={t("courseManager.unsavedChangesMessage")}
-        confirmAction={{
-          label: t("common.leave"),
-          onClick: handleReturnToCourses,
-        }}
-        cancelAction={{
-          label: t("common.stay"),
-          onClick: () => {
-            // Just close the dialog
-          },
-        }}
-      />
     </PageContainer>
   );
 };
